+++ src/main/java/ec/edu/ups/Biblioteca/model/Usuario.java
@@ -19,26 +19,26 @@ public class Usuario {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     private String username;
     private String password;
     private String rol;
     private String correo;
 
     // Getters and setters
-    public Long getId() {
-        return id;
-    }
+    public int getId() {
+		return id;
+	}
 
-    public void setId(Long id) {
-        this.id = id;
-    }
+	public void setId(int id) {
+		this.id = id;
+	}
 
     public String getUsername() {
         return username;
     }
 
-    public void setUsername(String username) {
+	public void setUsername(String username) {
         this.username = username;
     }
