+++ src/main/java/ec/edu/ups/Biblioteca/service/UsuarioService.java
@@ -29,7 +29,7 @@ public class UsuarioService  {
     @GET
     @Produces("application/json")
     @Path("/{id}")
-    public Usuario getUsuarioById(@PathParam("id") Long id) {
+    public Usuario getUsuarioById(@PathParam("id") int id) {
         return gestionUsuario.getUsuarioById(id);
     }
 
@@ -43,7 +43,7 @@ public class UsuarioService  {
     @PUT
     @Produces("application/json")
     @Path("/{id}")
-    public Response updateUsuario(@PathParam("id") Long id, Usuario usuario) {
+    public Response updateUsuario(@PathParam("id") int id, Usuario usuario) {
         usuario.setId(id);
         gestionUsuario.updateUsuario(usuario);
         return Response.ok().entity(usuario).build();
@@ -52,7 +52,7 @@ public class UsuarioService  {
     @DELETE
     @Produces("application/json")
     @Path("/{id}")
-    public Response deleteUsuario(@PathParam("id") Long id) {
+    public Response deleteUsuario(@PathParam("id") int id) {
         gestionUsuario.deleteUsuario(id);
         return Response.status(Response.Status.NO_CONTENT).build();
     }
